+++ contracts/modules/delegates/IexecAccessorsBoostDelegate.sol
@@ -18,8 +18,21 @@
 
 pragma solidity ^0.8.0;
 
-import "../../libs/IexecLibCore_v5.sol";
+import "../DelegateBase.v8.sol";
+import "../interfaces/IexecAccessorsBoost.sol";
 
-interface IexecAccessorsBoost {
-    function viewDealBoost(bytes32 id) external view returns (IexecLibCore_v5.DealBoost memory);
+/**
+ * @title Access data related to Poco Boost deals.
+ * @notice Access to PoCo Boost tasks must be done with classic `IexecAccessors`.
+ */
+contract IexecAccessorsBoostDelegate is IexecAccessorsBoost, DelegateBase {
+    /**
+     * Read deal Boost information.
+     * @param id The ID of the deal.
+     */
+    function viewDealBoost(
+        bytes32 id
+    ) external view returns (IexecLibCore_v5.DealBoost memory deal) {
+        return m_dealsBoost[id];
+    }
 }