+++ test/byContract/IexecPocoBoost/IexecPocoBoostDelegate.test.ts
@@ -6,7 +6,8 @@ import { assert, ethers } from 'hardhat';
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
 import { Contract, ContractFactory } from '@ethersproject/contracts';
 import {
-    IexecPocoBoostDelegate__factory,
+    IexecPocoBoostDelegateProxy__factory,
+    IexecAccessorsBoostDelegate__factory,
     IexecPocoBoostDelegate,
     App__factory,
     Workerpool__factory,
@@ -42,6 +43,7 @@ import {
     getTaskId,
     getDealId,
 } from '../../../utils/poco-tools';
+import { IexecLibCore_v5 } from '../../../typechain/contracts/modules/interfaces/IexecAccessorsBoost';
 
 chai.use(smock.matchers);
 
@@ -79,7 +81,7 @@ async function deployBoostFixture() {
     // Using native smock call here for understandability purposes (also works with
     // the custom `createMock` method)
     const iexecPocoBoostInstance = (await smock
-        .mock<IexecPocoBoostDelegate__factory>('IexecPocoBoostDelegate', {
+        .mock<IexecPocoBoostDelegateProxy__factory>('IexecPocoBoostDelegateProxy', {
             libraries: {
                 ['contracts/libs/IexecLibOrders_v5.sol:IexecLibOrders_v5']:
                     iexecLibOrdersInstanceAddress,
@@ -293,7 +295,7 @@ describe('IexecPocoBoostDelegate', function () {
             await expectOrderConsumed(iexecPocoBoostInstance, datasetOrderHash, expectedVolume);
             await expectOrderConsumed(iexecPocoBoostInstance, workerpoolOrderHash, expectedVolume);
             await expectOrderConsumed(iexecPocoBoostInstance, requestOrderHash, expectedVolume);
-            const deal = await iexecPocoBoostInstance.viewDealBoost(dealId);
+            const deal = await viewDealBoost(dealId);
             // Check addresses.
             expect(deal.requester).to.be.equal(requestOrder.requester, 'Requester mismatch');
             expect(deal.appOwner).to.be.equal(appProvider.address, 'App owner mismatch');
@@ -461,7 +463,7 @@ describe('IexecPocoBoostDelegate', function () {
                     volume,
                 );
             await expectOrderConsumed(iexecPocoBoostInstance, datasetOrderHash, undefined);
-            const deal = await iexecPocoBoostInstance.viewDealBoost(dealId);
+            const deal = await viewDealBoost(dealId);
             expect(deal.datasetPrice).to.be.equal(0);
         });
 
@@ -505,7 +507,7 @@ describe('IexecPocoBoostDelegate', function () {
             await expectOrderConsumed(iexecPocoBoostInstance, datasetOrderHash, expectedVolume);
             await expectOrderConsumed(iexecPocoBoostInstance, workerpoolOrderHash, expectedVolume);
             await expectOrderConsumed(iexecPocoBoostInstance, requestOrderHash, expectedVolume);
-            const deal = await iexecPocoBoostInstance.viewDealBoost(dealId);
+            const deal = await viewDealBoost(dealId);
             expect(deal.botFirst).to.be.equal(0);
             expect(deal.botSize).to.be.equal(expectedVolume);
         });
@@ -550,7 +552,7 @@ describe('IexecPocoBoostDelegate', function () {
             await expectOrderConsumed(iexecPocoBoostInstance, datasetOrderHash, expectedVolume);
             await expectOrderConsumed(iexecPocoBoostInstance, workerpoolOrderHash, expectedVolume);
             await expectOrderConsumed(iexecPocoBoostInstance, requestOrderHash, expectedVolume);
-            const deal = await iexecPocoBoostInstance.viewDealBoost(dealId);
+            const deal = await viewDealBoost(dealId);
             expect(deal.botFirst).to.be.equal(0);
             expect(deal.botSize).to.be.equal(expectedVolume);
         });
@@ -595,7 +597,7 @@ describe('IexecPocoBoostDelegate', function () {
             await expectOrderConsumed(iexecPocoBoostInstance, datasetOrderHash, expectedVolume);
             await expectOrderConsumed(iexecPocoBoostInstance, workerpoolOrderHash, expectedVolume);
             await expectOrderConsumed(iexecPocoBoostInstance, requestOrderHash, expectedVolume);
-            const deal = await iexecPocoBoostInstance.viewDealBoost(dealId);
+            const deal = await viewDealBoost(dealId);
             expect(deal.botFirst).to.be.equal(0);
             expect(deal.botSize).to.be.equal(expectedVolume);
         });
@@ -635,7 +637,7 @@ describe('IexecPocoBoostDelegate', function () {
                     requestOrderHash,
                     3,
                 );
-            const deal = await iexecPocoBoostInstance.viewDealBoost(dealId1);
+            const deal = await viewDealBoost(dealId1);
             expect(deal.botFirst).to.be.equal(0);
             expect(deal.botSize).to.be.equal(3);
 
@@ -663,7 +665,7 @@ describe('IexecPocoBoostDelegate', function () {
                     requestOrderHash,
                     5,
                 );
-            const deal2 = await iexecPocoBoostInstance.viewDealBoost(dealId2);
+            const deal2 = await viewDealBoost(dealId2);
             expect(deal2.botFirst).to.be.equal(3); // next index after last task of deal1:{0, 1, 2}
             expect(deal2.botSize).to.be.equal(5);
             // Verify request is fully consumed
@@ -1393,9 +1395,7 @@ describe('IexecPocoBoostDelegate', function () {
                 appProvider.address,
                 initialAppOwnerBalance,
             );
-            const expectedWorkerReward = (
-                await iexecPocoBoostInstance.viewDealBoost(dealId)
-            ).workerReward.toNumber();
+            const expectedWorkerReward = (await viewDealBoost(dealId)).workerReward.toNumber();
             // Worker reward formula already checked in match orders test, hence
             // we just need to verify here that some worker reward value will be
             // transferred
@@ -1947,6 +1947,16 @@ describe('IexecPocoBoostDelegate', function () {
             ).to.be.revertedWith('PocoBoost: Deadline not reached');
         });
     });
+
+    /**
+     * @notice Smock does not support getting struct variable from a mapping.
+     */
+    async function viewDealBoost(dealId: string) {
+        return await IexecAccessorsBoostDelegate__factory.connect(
+            iexecPocoBoostInstance.address,
+            anyone,
+        ).viewDealBoost(dealId);
+    }
 });
 
 /**