+++ CHANGELOG.md
@@ -50,6 +50,7 @@
     - Handle callback. (#183)
 - Implement `claimBoost`:
     - Refund requester. (#198, #201)
+- Add dedicated Poco Boost accessors.
 - Refactor tests. (#177, #178, #179, #181, #188, #200)
 - Add mocha to use test explorer. (#180)